initial release commit

+++ src/main/java/ru/kontur/cajrr/api/Table.java
@@ -39,7 +39,7 @@ public class Table implements Comparable<Table>{
     }
 
     @JsonProperty
-    public long getSlices() {
+    public int getSlices() {
         return this.slices;
     }
